Updated Plant System

+++ vrp/server/classes/Inventory/Growable.lua
@@ -7,7 +7,7 @@
 -- ****************************************************************************
 Growable = inherit(Object)
 
-function Growable:constructor(id, type, typeData, pos, ownerId, size, planted, lastGrown, lastWatered)
+function Growable:constructor(id, type, typeData, pos, ownerId, size, planted, lastGrown, lastWatered, timesEarned)
 	self.m_Id = id
 	self.m_Type = type
 	self.m_Object = createObject(typeData["Object"], pos)
@@ -17,6 +17,7 @@ function Growable:constructor(id, type, typeData, pos, ownerId, size, planted, l
 	self.m_LastGrown = lastGrown
 	self.m_LastWatered = lastWatered
 	self.m_OwnerId = ownerId
+	self.m_TimesEarned = timesEarned or 0
 
 	self.ms_GrowPerHour = typeData["GrowPerHour"]
 	self.ms_GrowPerHourWatered = typeData["GrowPerHourWatered"]
@@ -26,6 +27,8 @@ function Growable:constructor(id, type, typeData, pos, ownerId, size, planted, l
 	self.ms_ItemPerSize = typeData["ItemPerSize"]
 	self.ms_ObjectSizeMin = typeData["ObjectSizeMin"]
 	self.ms_ObjectSizeSteps = typeData["ObjectSizeSteps"]
+	self.ms_TimesEarnedForDestroy = typeData["TimesEarnedForDestroy"]
+	self.ms_Illegal = typeData["Illegal"]
 
 	self.m_Colshape = createColSphere(pos.x, pos.y, pos.z+1, 1)
 	addEventHandler("onColShapeHit", self.m_Colshape, bind(self.onColShapeHit, self))
@@ -76,7 +79,7 @@ function Growable:harvest(player)
 	if not player.vehicle then
 	--if player:getId() == self.m_OwnerId or (player:getFaction() and player:getFaction():isStateFaction() and player:isFactionDuty()) then
 		local amount = self.m_Size*self.ms_ItemPerSize
-		if player:getFaction() and player:getFaction():isStateFaction() and player:isFactionDuty() then
+		if self.ms_Illegal and player:getFaction() and player:getFaction():isStateFaction() and player:isFactionDuty() then
 			player:sendInfo(_("Du hast %d %s sichergestellt!", player, amount, self.ms_Item))
 			player:getFaction():giveMoney(amount*5, "Drogen-Asservation")
 			player:triggerEvent("hidePlantGUI")
@@ -90,9 +93,13 @@ function Growable:harvest(player)
 				player:getInventory():giveItem(self.ms_Item, amount)
 				player:triggerEvent("hidePlantGUI")
 				self.m_Size = 0
-				sql:queryExec("DELETE FROM ??_plants WHERE Id = ?", sql:getPrefix(), self.m_Id)
+				self.m_TimesEarned = self.m_TimesEarned + 1
+				self:refreshObjectSize()
 				StatisticsLogger:getSingleton():addDrugHarvestLog(player, self.m_Type, self.m_OwnerId, amount, 0)
-				delete(self)
+				if self.m_TimesEarned >= self.ms_TimesEarnedForDestroy  then
+					sql:queryExec("DELETE FROM ??_plants WHERE Id = ?", sql:getPrefix(), self.m_Id)
+					delete(self)
+				end
 			else
 				player:sendError(_("Du hast in deinem Inventar nicht Platz für %d %s!", player, amount, self.ms_Item))
 			end
@@ -124,7 +131,7 @@ function Growable:waterPlant(player)
 end
 
 function Growable:save()
-	local result = sql:queryExec("UPDATE ??_plants SET Size = ?, last_grown = ?, last_watered = ? WHERE Id = ?", sql:getPrefix(), self.m_Size, self.m_LastGrown, self.m_LastWatered, self.m_Id)
+	local result = sql:queryExec("UPDATE ??_plants SET Size = ?, last_grown = ?, last_watered = ?, times_earned = ? WHERE Id = ?", sql:getPrefix(), self.m_Size, self.m_LastGrown, self.m_LastWatered, self.m_TimesEarned, self.m_Id)
 	if not result then outputDebug("Plant ID "..self.m_Id.." not saved!") end
 end
